LayoutTests/fast/css/border-shorthand-initialize-longhands-expected.txt

This test ensures border shorthand property initializes longhand properties such as border-top-style and border-right-color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL div.setAttribute('style', 'border: 1px');div.style.borderTopStyle should be none. Was initial.
FAIL div.style.borderRightStyle should be none. Was initial.
FAIL div.style.borderBottomStyle should be none. Was initial.
FAIL div.style.borderLeftStyle should be none. Was initial.
FAIL div.style.borderTopColor should be currentcolor. Was initial.
FAIL div.style.borderRightColor should be currentcolor. Was initial.
FAIL div.style.borderBottomColor should be currentcolor. Was initial.
FAIL div.style.borderLeftColor should be currentcolor. Was initial.
FAIL div.setAttribute('style', 'border: solid');div.style.borderTopWidth should be medium. Was initial.
FAIL div.style.borderRightWidth should be medium. Was initial.
FAIL div.style.borderBottomWidth should be medium. Was initial.
FAIL div.style.borderLeftWidth should be medium. Was initial.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
